added query by user_id where not yet used

+++ database/delete.php
@@ -2,9 +2,9 @@
 require_once '../Usefull-PHP/up_database.php';
 
 class Delete{
-    public static function deleteHomework($id){
-        $stmt = up_database::prepare("DELETE FROM homework WHERE id = ?");
-        $stmt->bind_param('s', $id);
+    public static function deleteHomework($id, $user_id){
+        $stmt = up_database::prepare("DELETE FROM homework WHERE id = ? and user_id = ?");
+        $stmt->bind_param('is', $id, $user_id);
         $stmt->execute();
         if($stmt->error != null){
             // todo do error logging
@@ -13,21 +13,9 @@ class Delete{
 
     }
 
-    public static function deleteCategory($id){
-        $stmt = up_database::prepare("DELETE FROM category WHERE id = ?");
-        $stmt->bind_param('s', $id);
-        $stmt->execute();
-        if($stmt->error != null){
-            // todo do error logging
-        }
-        $stmt->close();
-
-        return json_encode(['status' => 'ok']);
-    }
-
-    public static function deleteUser($id){
-        $stmt = up_database::prepare("DELETE FROM user WHERE id = ?");
-        $stmt->bind_param('s', $id);
+    public static function deleteCategory($id, $user_id){
+        $stmt = up_database::prepare("DELETE FROM category WHERE id = ? and user_id = ?");
+        $stmt->bind_param('is', $id, $user_id);
         $stmt->execute();
         if($stmt->error != null){
             // todo do error logging
@@ -36,5 +24,17 @@ class Delete{
 
         return json_encode(['status' => 'ok']);
     }
+
+    // public static function deleteUser($id){
+    //     $stmt = up_database::prepare("DELETE FROM user WHERE id = ?");
+    //     $stmt->bind_param('s', $id);
+    //     $stmt->execute();
+    //     if($stmt->error != null){
+    //         // todo do error logging
+    //     }
+    //     $stmt->close();
+
+    //     return json_encode(['status' => 'ok']);
+    // }
 }
 ?>
